исправлены критические замечания
