+++ MovieQuiz/Presentation/AlertPresenter.swift
@@ -8,7 +8,6 @@ final class AlertPresenter {
             message: model.message,
             preferredStyle: .alert
         )
-        alert.view.accessibilityIdentifier = "Alert"
         
         let action = UIAlertAction(title: model.buttonText, style: .default) { _ in
             model.completion?()
