+++ MovieQuiz/Presentation/MovieQuizViewController.swift
@@ -11,7 +11,7 @@ final class MovieQuizViewController: UIViewController, MovieQuizViewControllerPr
     @IBOutlet private var noButton: UIButton!
     
     // MARK: - Private Properties
-    private var presenter: MovieQuizPresenter!
+    private var presenter: MovieQuizPresenter?
     
     // MARK: - Lifecycle
     override func viewDidLoad() {
@@ -25,11 +25,11 @@ final class MovieQuizViewController: UIViewController, MovieQuizViewControllerPr
     
     // MARK: - IBActions
     @IBAction private func yesButtonClicked(_ sender: UIButton) {
-        presenter.yesButtonClicked()
+        presenter?.buttonClicked(isYes: true)
     }
     
     @IBAction private func noButtonClicked(_ sender: UIButton) {
-        presenter.noButtonClicked()
+        presenter?.buttonClicked(isYes: false)
     }
     
     // MARK: - Private functions
@@ -42,17 +42,19 @@ final class MovieQuizViewController: UIViewController, MovieQuizViewControllerPr
     }
     
     func show(quiz result: QuizResultsViewModel) {
-        let message = presenter.makeResultsMessage()
+        let message = presenter?.makeResultsMessage()
         
         let alert = UIAlertController(
             title: result.title,
             message: message,
             preferredStyle: .alert)
         
+        alert.view.accessibilityIdentifier = "Game results"
+        
         let action = UIAlertAction(title: result.buttonText, style: .default) { [weak self] _ in
             guard let self = self else { return }
             
-            self.presenter.restartGame()
+            self.presenter?.restartGame()
         }
         
         alert.addAction(action)
@@ -87,7 +89,7 @@ final class MovieQuizViewController: UIViewController, MovieQuizViewControllerPr
                                    style: .default) { [weak self] _ in
             guard let self = self else { return }
             
-            self.presenter.restartGame()
+            self.presenter?.restartGame()
         }
         
         alert.addAction(action)
